refactor(doctor): extract streaming alias migration helpers
