+++ src/commands/doctor-legacy-config.ts
@@ -1,14 +1,7 @@
 import { shouldMoveSingleAccountChannelKey } from "../channels/plugins/setup-helpers.js";
 import type { OpenClawConfig } from "../config/config.js";
-import {
-  formatSlackStreamingBooleanMigrationMessage,
-  formatSlackStreamModeMigrationMessage,
-  resolveDiscordPreviewStreamMode,
-  resolveSlackNativeStreaming,
-  resolveSlackStreamingMode,
-  resolveTelegramPreviewStreamMode,
-} from "../config/discord-preview-streaming.js";
 import { DEFAULT_ACCOUNT_ID } from "../routing/session-key.js";
+import { normalizeStreamingAliasesForProvider } from "./doctor-legacy-streaming-aliases.js";
 
 export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
   config: OpenClawConfig;
@@ -101,122 +94,6 @@ export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
     return { entry: updated, changed };
   };
 
-  const normalizePreviewStreamingAliases = (params: {
-    entry: Record<string, unknown>;
-    pathPrefix: string;
-    resolveStreaming: (entry: Record<string, unknown>) => string;
-  }): { entry: Record<string, unknown>; changed: boolean } => {
-    let updated = params.entry;
-    const hadLegacyStreamMode = updated.streamMode !== undefined;
-    const beforeStreaming = updated.streaming;
-    const resolved = params.resolveStreaming(updated);
-    const shouldNormalize =
-      hadLegacyStreamMode ||
-      typeof beforeStreaming === "boolean" ||
-      (typeof beforeStreaming === "string" && beforeStreaming !== resolved);
-    if (!shouldNormalize) {
-      return { entry: updated, changed: false };
-    }
-
-    let changed = false;
-    if (beforeStreaming !== resolved) {
-      updated = { ...updated, streaming: resolved };
-      changed = true;
-    }
-    if (hadLegacyStreamMode) {
-      const { streamMode: _ignored, ...rest } = updated;
-      updated = rest;
-      changed = true;
-      changes.push(
-        `Moved ${params.pathPrefix}.streamMode → ${params.pathPrefix}.streaming (${resolved}).`,
-      );
-    }
-    if (typeof beforeStreaming === "boolean") {
-      changes.push(`Normalized ${params.pathPrefix}.streaming boolean → enum (${resolved}).`);
-    } else if (typeof beforeStreaming === "string" && beforeStreaming !== resolved) {
-      changes.push(
-        `Normalized ${params.pathPrefix}.streaming (${beforeStreaming}) → (${resolved}).`,
-      );
-    }
-
-    return { entry: updated, changed };
-  };
-
-  const normalizeSlackStreamingAliases = (params: {
-    entry: Record<string, unknown>;
-    pathPrefix: string;
-  }): { entry: Record<string, unknown>; changed: boolean } => {
-    let updated = params.entry;
-    const hadLegacyStreamMode = updated.streamMode !== undefined;
-    const legacyStreaming = updated.streaming;
-    const beforeStreaming = updated.streaming;
-    const beforeNativeStreaming = updated.nativeStreaming;
-    const resolvedStreaming = resolveSlackStreamingMode(updated);
-    const resolvedNativeStreaming = resolveSlackNativeStreaming(updated);
-    const shouldNormalize =
-      hadLegacyStreamMode ||
-      typeof legacyStreaming === "boolean" ||
-      (typeof legacyStreaming === "string" && legacyStreaming !== resolvedStreaming);
-    if (!shouldNormalize) {
-      return { entry: updated, changed: false };
-    }
-
-    let changed = false;
-    if (beforeStreaming !== resolvedStreaming) {
-      updated = { ...updated, streaming: resolvedStreaming };
-      changed = true;
-    }
-    if (
-      typeof beforeNativeStreaming !== "boolean" ||
-      beforeNativeStreaming !== resolvedNativeStreaming
-    ) {
-      updated = { ...updated, nativeStreaming: resolvedNativeStreaming };
-      changed = true;
-    }
-    if (hadLegacyStreamMode) {
-      const { streamMode: _ignored, ...rest } = updated;
-      updated = rest;
-      changed = true;
-      changes.push(formatSlackStreamModeMigrationMessage(params.pathPrefix, resolvedStreaming));
-    }
-    if (typeof legacyStreaming === "boolean") {
-      changes.push(
-        formatSlackStreamingBooleanMigrationMessage(params.pathPrefix, resolvedNativeStreaming),
-      );
-    } else if (typeof legacyStreaming === "string" && legacyStreaming !== resolvedStreaming) {
-      changes.push(
-        `Normalized ${params.pathPrefix}.streaming (${legacyStreaming}) → (${resolvedStreaming}).`,
-      );
-    }
-
-    return { entry: updated, changed };
-  };
-
-  const normalizeStreamingAliasesForProvider = (params: {
-    provider: "telegram" | "slack" | "discord";
-    entry: Record<string, unknown>;
-    pathPrefix: string;
-  }): { entry: Record<string, unknown>; changed: boolean } => {
-    if (params.provider === "telegram") {
-      return normalizePreviewStreamingAliases({
-        entry: params.entry,
-        pathPrefix: params.pathPrefix,
-        resolveStreaming: resolveTelegramPreviewStreamMode,
-      });
-    }
-    if (params.provider === "discord") {
-      return normalizePreviewStreamingAliases({
-        entry: params.entry,
-        pathPrefix: params.pathPrefix,
-        resolveStreaming: resolveDiscordPreviewStreamMode,
-      });
-    }
-    return normalizeSlackStreamingAliases({
-      entry: params.entry,
-      pathPrefix: params.pathPrefix,
-    });
-  };
-
   const normalizeProvider = (provider: "telegram" | "slack" | "discord") => {
     const channels = next.channels as Record<string, unknown> | undefined;
     const rawEntry = channels?.[provider];
@@ -242,6 +119,7 @@ export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
     });
     updated = providerStreaming.entry;
     changed = changed || providerStreaming.changed;
+    changes.push(...providerStreaming.changes);
 
     const rawAccounts = updated.accounts;
     if (isRecord(rawAccounts)) {
@@ -269,6 +147,7 @@ export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
         });
         accountEntry = accountStreaming.entry;
         accountChanged = accountChanged || accountStreaming.changed;
+        changes.push(...accountStreaming.changes);
         if (accountChanged) {
           accounts[accountId] = accountEntry;
           accountsChanged = true;
